+++ src/cli/domain/watch.js
@@ -3,6 +3,8 @@
 const path = require('path');
 const watch = require('watch');
 
+const settings = require('../../resources/settings');
+
 module.exports = function(dirs, baseDir, changed) {
   try {
     watch.watchTree(
@@ -12,13 +14,13 @@ module.exports = function(dirs, baseDir, changed) {
         ignoreUnreadableDir: true,
         ignoreDotFiles: false,
         filter: fileOrDir =>
-          /node_modules|package\.tar\.gz|_package|\.sw[op]|\.git|\.DS_Store|oc\.json/.test(
-            fileOrDir
-          ) === false
+          settings.filesToIgnoreOnDevWatch.test(fileOrDir) === false
       },
       (fileName, currentStat, previousStat) => {
         if (!!currentStat || !!previousStat) {
-          const componentDir = dirs.find(dir => Boolean(fileName.match(dir)));
+          const componentDir = dirs.find(dir =>
+            Boolean(fileName.match(dir + path.sep))
+          );
           changed(null, fileName, componentDir);
         }
       }