+++ src/registry/app-start.js
@@ -17,8 +17,8 @@ module.exports = function(repository, options, callback) {
   logger.log(
     format(
       colors.yellow('Connecting to library: {0}/{1}'),
-      options.s3.bucket,
-      options.s3.componentsDir
+      options.storage.options.bucket,
+      options.storage.options.componentsDir
     )
   );
 
@@ -54,9 +54,7 @@ module.exports = function(repository, options, callback) {
             logger.log(colors.green('Component published.'));
           } else {
             logger.log(
-              colors.red(
-                format('Component not published: {0}', _.first(err).message)
-              )
+              colors.red(format('Component not published: {0}', err.message))
             );
           }