+++ src/registry/domain/components-cache/components-list.js
@@ -7,7 +7,8 @@ const _ = require('lodash');
 const getUnixUTCTimestamp = require('oc-get-unix-utc-timestamp');
 
 module.exports = (conf, cdn) => {
-  const filePath = () => `${conf.s3.componentsDir}/components.json`;
+  const filePath = () =>
+    `${conf.storage.options.componentsDir}/components.json`;
 
   const componentsList = {
     getFromJson: callback => cdn.getJson(filePath(), true, callback),
@@ -17,7 +18,7 @@ module.exports = (conf, cdn) => {
 
       const getVersionsForComponent = (componentName, cb) => {
         cdn.listSubDirectories(
-          `${conf.s3.componentsDir}/${componentName}`,
+          `${conf.storage.options.componentsDir}/${componentName}`,
           (err, versions) => {
             if (err) {
               return cb(err);
@@ -27,33 +28,36 @@ module.exports = (conf, cdn) => {
         );
       };
 
-      cdn.listSubDirectories(conf.s3.componentsDir, (err, components) => {
-        if (err) {
-          if (err.code === 'dir_not_found') {
-            return callback(null, {
-              lastEdit: getUnixUTCTimestamp(),
-              components: []
+      cdn.listSubDirectories(
+        conf.storage.options.componentsDir,
+        (err, components) => {
+          if (err) {
+            if (err.code === 'dir_not_found') {
+              return callback(null, {
+                lastEdit: getUnixUTCTimestamp(),
+                components: []
+              });
+            }
+
+            return callback(err);
+          }
+
+          async.map(components, getVersionsForComponent, (errors, versions) => {
+            if (errors) {
+              return callback(errors);
+            }
+
+            _.forEach(components, (component, i) => {
+              componentsInfo[component] = versions[i];
             });
-          }
 
-          return callback(err);
+            callback(null, {
+              lastEdit: getUnixUTCTimestamp(),
+              components: componentsInfo
+            });
+          });
         }
-
-        async.map(components, getVersionsForComponent, (errors, versions) => {
-          if (errors) {
-            return callback(errors);
-          }
-
-          _.forEach(components, (component, i) => {
-            componentsInfo[component] = versions[i];
-          });
-
-          callback(null, {
-            lastEdit: getUnixUTCTimestamp(),
-            components: componentsInfo
-          });
-        });
-      });
+      );
     },
 
     refresh: callback => {