+++ src/registry/domain/s3.js
@@ -11,21 +11,20 @@ const _ = require('lodash');
 const getFileInfo = require('../../utils/get-file-info');
 const getNextYear = require('../../utils/get-next-year');
 const strings = require('../../resources');
-
 module.exports = function(conf) {
-  const httpOptions = { timeout: conf.s3.timeout || 10000 };
-  if (conf.s3.agentProxy) {
-    httpOptions.agent = conf.s3.agentProxy;
+  const httpOptions = { timeout: conf.timeout || 10000 };
+  if (conf.agentProxy) {
+    httpOptions.agent = conf.agentProxy;
   }
 
   AWS.config.update({
-    accessKeyId: conf.s3.key,
-    secretAccessKey: conf.s3.secret,
-    region: conf.s3.region,
+    accessKeyId: conf.key,
+    secretAccessKey: conf.secret,
+    region: conf.region,
     httpOptions
   });
 
-  const bucket = conf.s3.bucket;
+  const bucket = conf.bucket;
   const cache = new Cache({
     verbose: !!conf.verbosity,
     refreshInterval: conf.refreshInterval
@@ -50,8 +49,8 @@ module.exports = function(conf) {
             return callback(
               err.code === 'NoSuchKey'
                 ? {
-                  code: strings.errors.s3.FILE_NOT_FOUND_CODE,
-                  msg: format(strings.errors.s3.FILE_NOT_FOUND, filePath)
+                  code: strings.errors.STORAGE.FILE_NOT_FOUND_CODE,
+                  msg: format(strings.errors.STORAGE.FILE_NOT_FOUND, filePath)
                 }
                 : err
             );
@@ -97,15 +96,15 @@ module.exports = function(conf) {
         callback(null, JSON.parse(file));
       } catch (er) {
         return callback({
-          code: strings.errors.s3.FILE_NOT_VALID_CODE,
-          msg: format(strings.errors.s3.FILE_NOT_VALID, filePath)
+          code: strings.errors.STORAGE.FILE_NOT_VALID_CODE,
+          msg: format(strings.errors.STORAGE.FILE_NOT_VALID, filePath)
         });
       }
     });
   };
 
   const getUrl = (componentName, version, fileName) =>
-    `${conf.s3.path}${componentName}/${version}/${fileName}`;
+    `${conf.path}${componentName}/${version}/${fileName}`;
 
   const listSubDirectories = (dir, callback) => {
     const normalisedPath =
@@ -126,8 +125,8 @@ module.exports = function(conf) {
 
         if (data.CommonPrefixes.length === 0) {
           return callback({
-            code: strings.errors.s3.DIR_NOT_FOUND_CODE,
-            msg: format(strings.errors.s3.DIR_NOT_FOUND, dir)
+            code: strings.errors.STORAGE.DIR_NOT_FOUND_CODE,
+            msg: format(strings.errors.STORAGE.DIR_NOT_FOUND, dir)
           });
         }
 
@@ -197,6 +196,7 @@ module.exports = function(conf) {
     maxConcurrentRequests: 20,
     putDir,
     putFile,
-    putFileContent
+    putFileContent,
+    adapterType: 's3'
   };
 };