+++ src/resources/settings.js
@@ -4,6 +4,7 @@ module.exports = {
   configFile: {
     src: './oc.json'
   },
+  filesToIgnoreOnDevWatch: /node_modules|package\.tar\.gz|_package|\.sw[op]|\.git\/|\.DS_Store|oc\.json/,
   maxLoopIterations: 1e9,
   registry: {
     acceptRenderedHeader: 'application/vnd.oc.rendered+json',